api/requirements.txt
fastapi
uvicorn
sqlmodel
pydantic
pytest
httpx
python-dotenv 